extract filter functions to use-cases
