+++ frontend/src/features/Dashboard/useCases/filters/filterByUnits.ts
@@ -1,0 +1,5 @@
+import type { Dashboard } from '@features/Dashboard/types'
+
+export function filterByUnits(controlUnits: number[], dashboard: Dashboard.DashboardFromApi): boolean {
+  return controlUnits.length === 0 || dashboard.controlUnitIds.some(control => controlUnits.includes(control))
+}
